backend/requirements.txt
openai>=1.3.0
reportlab>=3.6.12
matplotlib>=3.6.3
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-ai
pdfplumber>=0.6.0
aiofiles>=23.1.0
httpx>=0.23.0
opentelemetry-api==1.8.0
opentelemetry-sdk==1.8.0
asyncio
logfire
